refactor(post-like.service): post-like 추가/삭제 로직 수정
요청된 자원의 추가/삭제가 이뤄지기 전 해당 자원의 존재여부를 확인하고
DB 작업을 진행 후에 로그를 출력하도록 수정.

+++ personal-blog-service/src/blog/service/post-like.serivce.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { ConflictException, Inject, Injectable } from '@nestjs/common';
 import { UserInfoService } from '../../user/service/user-info.service';
 import { PostLikeDto } from '../dto/post-like.dto';
 import { WINSTON_MODULE_PROVIDER } from 'nest-winston';
@@ -23,19 +23,33 @@ export class PostLikeService {
   }
 
   async addPostLikeUser(postLikeDto: PostLikeDto): Promise<void> {
-    this.logger.info(`addPostLikeUser - [${JSON.stringify(postLikeDto)}]`);
+    const isExist = this.postLikeRepository.isExist(
+      PostLikeDao.from({ ...postLikeDto }).toPostLikeEntity(),
+    );
+    if (isExist) {
+      throw new ConflictException('PostId and uid is already exist!');
+    }
 
     await this.postLikeRepository.savePostLikeEntity(
       PostLikeDao.from({ ...postLikeDto }).toPostLikeEntity(),
     );
+
+    this.logger.info(`addPostLikeUser - [${JSON.stringify(postLikeDto)}]`);
   }
 
   async removePostLikeUser(postLikeDto: PostLikeDto): Promise<void> {
-    this.logger.info(`removePostLikeUser - [${JSON.stringify(postLikeDto)}]`);
+    const isExist = this.postLikeRepository.isExist(
+      PostLikeDao.from({ ...postLikeDto }).toPostLikeEntity(),
+    );
+    if (!isExist) {
+      throw new ConflictException('PostId and uid does not exist!');
+    }
 
     await this.postLikeRepository.removePostLikeDto(
       PostLikeDao.from({ ...postLikeDto }).toPostLikeEntity(),
     );
+
+    this.logger.info(`removePostLikeUser - [${JSON.stringify(postLikeDto)}]`);
   }
 
   private async getNicknameList(
